requirements.txt
# Core Dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0

# LangChain
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain>=0.0.300

# Document Processing
python-pptx==0.6.21
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.2
beautifulsoup4==4.12.2

# Vector Database
faiss-cpu>=1.11.0

# Hugging Face
huggingface-hub>=0.16.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.21.0  # For better performance on CPU/GPU

# Web Interface
streamlit==1.27.2
streamlit-chat==0.1.1
